Updated to C++23 Fixed implicit capture in lambdas. Disable a macro in Globals.h that overrides the new operator because it causes builds to fail on MSVC. Updated FMT lib. Fixed broken api changes. Added format_as function for FMT because enums are no longer implicitly converted to ints.

+++ src/Entities/ProjectileEntity.cpp
@@ -286,7 +286,7 @@ std::unique_ptr<cProjectileEntity> cProjectileEntity::Create(
 		}
 	}
 
-	LOGWARNING("%s: Unknown projectile kind: %d", __FUNCTION__, a_Kind);
+	LOGWARNING("%s: Unknown projectile kind: %d", __FUNCTION__, static_cast<int>(a_Kind));
 	return nullptr;
 }
